add unit test for plugin settings and list plugin feature/config
