--- notifications/core/src/test/kotlin/org/opensearch/notifications/core/ChimeDestinationTests.kt
@@ -1,212 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- *
- * Modifications Copyright OpenSearch Contributors. See
- * GitHub history for details.
- */
-
-/*
- * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- *
- */
-
-package org.opensearch.notifications.core
-
-import org.apache.http.client.methods.CloseableHttpResponse
-import org.apache.http.client.methods.HttpPost
-import org.apache.http.entity.StringEntity
-import org.apache.http.impl.client.CloseableHttpClient
-import org.apache.http.message.BasicStatusLine
-import org.easymock.EasyMock
-import org.junit.jupiter.api.Assertions
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.Arguments
-import org.junit.jupiter.params.provider.MethodSource
-import org.opensearch.notifications.core.client.DestinationHttpClient
-import org.opensearch.notifications.core.transport.DestinationTransportProvider
-import org.opensearch.notifications.core.transport.WebhookDestinationTransport
-import org.opensearch.notifications.spi.model.DestinationMessageResponse
-import org.opensearch.notifications.spi.model.MessageContent
-import org.opensearch.notifications.spi.model.destination.ChimeDestination
-import org.opensearch.notifications.spi.model.destination.DestinationType
-import org.opensearch.rest.RestStatus
-import java.net.MalformedURLException
-import java.util.stream.Stream
-
-internal class ChimeDestinationTests {
-    companion object {
-        @JvmStatic
-        fun escapeSequenceToRaw(): Stream<Arguments> =
-            Stream.of(
-                Arguments.of("\n", """\n"""),
-                Arguments.of("\t", """\t"""),
-                Arguments.of("\b", """\b"""),
-                Arguments.of("\r", """\r"""),
-                Arguments.of("\"", """\""""),
-            )
-    }
-
-    @Test
-    fun `test chime message null entity response`() {
-        val mockHttpClient: CloseableHttpClient = EasyMock.createMock(CloseableHttpClient::class.java)
-
-        // The DestinationHttpClient replaces a null entity with "{}".
-        val expectedWebhookResponse = DestinationMessageResponse(RestStatus.OK.status, "{}")
-        // TODO replace EasyMock in all UTs with mockk which fits Kotlin better
-        val httpResponse: CloseableHttpResponse = EasyMock.createMock(CloseableHttpResponse::class.java)
-        EasyMock.expect(mockHttpClient.execute(EasyMock.anyObject(HttpPost::class.java))).andReturn(httpResponse)
-
-        val mockStatusLine: BasicStatusLine = EasyMock.createMock(BasicStatusLine::class.java)
-        EasyMock.expect(httpResponse.statusLine).andReturn(mockStatusLine)
-        EasyMock.expect(httpResponse.entity).andReturn(null).anyTimes()
-        EasyMock.expect(mockStatusLine.statusCode).andReturn(RestStatus.OK.status)
-        EasyMock.replay(mockHttpClient)
-        EasyMock.replay(httpResponse)
-        EasyMock.replay(mockStatusLine)
-
-        val httpClient = DestinationHttpClient(mockHttpClient)
-        val webhookDestinationTransport = WebhookDestinationTransport(httpClient)
-        DestinationTransportProvider.destinationTransportMap = mapOf(DestinationType.CHIME to webhookDestinationTransport)
-
-        val title = "test Chime"
-        val messageText = "Message gughjhjlkh Body emoji test: :) :+1: " +
-            "link test: http://sample.com email test: marymajor@example.com All member call out: " +
-            "@All All Present member call out: @Present"
-        val url = "https://abc/com"
-
-        val destination = ChimeDestination(url)
-        val message = MessageContent(title, messageText)
-
-        val actualChimeResponse: DestinationMessageResponse = NotificationCoreImpl.sendMessage(destination, message, "referenceId")
-
-        assertEquals(expectedWebhookResponse.statusText, actualChimeResponse.statusText)
-        assertEquals(expectedWebhookResponse.statusCode, actualChimeResponse.statusCode)
-    }
-
-    @Test
-    fun `test chime message empty entity response`() {
-        val mockHttpClient: CloseableHttpClient = EasyMock.createMock(CloseableHttpClient::class.java)
-        val expectedWebhookResponse = DestinationMessageResponse(RestStatus.OK.status, "")
-
-        val httpResponse: CloseableHttpResponse = EasyMock.createMock(CloseableHttpResponse::class.java)
-        EasyMock.expect(mockHttpClient.execute(EasyMock.anyObject(HttpPost::class.java))).andReturn(httpResponse)
-        val mockStatusLine: BasicStatusLine = EasyMock.createMock(BasicStatusLine::class.java)
-        EasyMock.expect(httpResponse.statusLine).andReturn(mockStatusLine)
-        EasyMock.expect(httpResponse.entity).andReturn(StringEntity("")).anyTimes()
-        EasyMock.expect(mockStatusLine.statusCode).andReturn(RestStatus.OK.status)
-        EasyMock.replay(mockHttpClient)
-        EasyMock.replay(httpResponse)
-        EasyMock.replay(mockStatusLine)
-
-        val httpClient = DestinationHttpClient(mockHttpClient)
-        val webhookDestinationTransport = WebhookDestinationTransport(httpClient)
-        DestinationTransportProvider.destinationTransportMap = mapOf(DestinationType.CHIME to webhookDestinationTransport)
-
-        val title = "test Chime"
-        val messageText = "{\"Content\":\"Message gughjhjlkh Body emoji test: :) :+1: " +
-            "link test: http://sample.com email test: marymajor@example.com All member call out: " +
-            "@All All Present member call out: @Present\"}"
-        val url = "https://abc/com"
-
-        val destination = ChimeDestination(url)
-        val message = MessageContent(title, messageText)
-
-        val actualChimeResponse: DestinationMessageResponse = NotificationCoreImpl.sendMessage(destination, message, "referenceId")
-
-        assertEquals(expectedWebhookResponse.statusText, actualChimeResponse.statusText)
-        assertEquals(expectedWebhookResponse.statusCode, actualChimeResponse.statusCode)
-    }
-
-    @Test
-    fun `test chime message non-empty entity response`() {
-        val responseContent = "It worked!"
-        val mockHttpClient: CloseableHttpClient = EasyMock.createMock(CloseableHttpClient::class.java)
-        val expectedWebhookResponse = DestinationMessageResponse(RestStatus.OK.status, responseContent)
-
-        val httpResponse: CloseableHttpResponse = EasyMock.createMock(CloseableHttpResponse::class.java)
-        EasyMock.expect(mockHttpClient.execute(EasyMock.anyObject(HttpPost::class.java))).andReturn(httpResponse)
-        val mockStatusLine: BasicStatusLine = EasyMock.createMock(BasicStatusLine::class.java)
-        EasyMock.expect(httpResponse.statusLine).andReturn(mockStatusLine)
-        EasyMock.expect(httpResponse.entity).andReturn(StringEntity(responseContent)).anyTimes()
-        EasyMock.expect(mockStatusLine.statusCode).andReturn(RestStatus.OK.status)
-        EasyMock.replay(mockHttpClient)
-        EasyMock.replay(httpResponse)
-        EasyMock.replay(mockStatusLine)
-
-        val httpClient = DestinationHttpClient(mockHttpClient)
-        val webhookDestinationTransport = WebhookDestinationTransport(httpClient)
-        DestinationTransportProvider.destinationTransportMap = mapOf(DestinationType.CHIME to webhookDestinationTransport)
-
-        val title = "test Chime"
-        val messageText = "{\"Content\":\"Message gughjhjlkh Body emoji test: :) :+1: " +
-            "link test: http://sample.com email test: marymajor@example.com All member call out: " +
-            "@All All Present member call out: @Present\"}"
-        val url = "https://abc/com"
-
-        val destination = ChimeDestination(url)
-        val message = MessageContent(title, messageText)
-
-        val actualChimeResponse: DestinationMessageResponse = NotificationCoreImpl.sendMessage(destination, message, "referenceId")
-
-        assertEquals(expectedWebhookResponse.statusText, actualChimeResponse.statusText)
-        assertEquals(expectedWebhookResponse.statusCode, actualChimeResponse.statusCode)
-    }
-
-    @Test
-    fun `test url missing should throw IllegalArgumentException with message`() {
-        val exception = Assertions.assertThrows(IllegalArgumentException::class.java) {
-            ChimeDestination("")
-        }
-        assertEquals("url is null or empty", exception.message)
-    }
-
-    @Test
-    fun testUrlInvalidMessage() {
-        assertThrows<MalformedURLException> {
-            ChimeDestination("invalidUrl")
-        }
-    }
-
-    @Test
-    fun `test content missing content should throw IllegalArgumentException`() {
-        val exception = Assertions.assertThrows(IllegalArgumentException::class.java) {
-            MessageContent("title", "")
-        }
-        assertEquals("text message part is null or empty", exception.message)
-    }
-
-    @ParameterizedTest
-    @MethodSource("escapeSequenceToRaw")
-    fun `test build request body for chime webhook should have title included and prevent escape`(
-        escapeSequence: String,
-        rawString: String
-    ) {
-        val httpClient = DestinationHttpClient()
-        val title = "test chime webhook"
-        val messageText = "line1${escapeSequence}line2"
-        val url = "https://abc/com"
-        val expectedRequestBody = """{"Content":"$title\n\nline1${rawString}line2"}"""
-        val destination = ChimeDestination(url)
-        val message = MessageContent(title, messageText)
-        val actualRequestBody = httpClient.buildRequestBody(destination, message)
-        assertEquals(expectedRequestBody, actualRequestBody)
-    }
-}
